back-end/workingrequirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
diskcache==5.6.3
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.14
MarkupSafe==3.0.2
numpy==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
